fix(web): move Node.js-only modules out of @jitaspace/utils
@jitaspace/utils is listed in transpilePackages in next.config.mjs, so
all its code is bundled into the browser build. Adding download.ts and
zip.ts there pulled in node-fetch and node-stream-zip (which require
Node.js 'fs' at the top level), breaking the Turbopack build with
"Module not found: Can't resolve 'fs'".

Move download.ts and zip.ts into @jitaspace/sde-utils, which is a
server/CLI-only package never transpiled for the browser. The CLI
wrappers in apps/cli/utils/{download,fs}.ts are updated to import from
@jitaspace/sde-utils accordingly, and the now-unnecessary @jitaspace/utils
dep is removed from apps/cli.
