+++ apps/cli/package.json
@@ -19,7 +19,6 @@
   "devDependencies": {
     "@asteasolutions/zod-to-openapi": "^4.8.0",
     "@jitaspace/sde-utils": "workspace:*",
-    "@jitaspace/utils": "workspace:*",
     "@commander-js/extra-typings": "^14.0.0",
     "@jitaspace/eslint-config": "workspace:*",
     "@jitaspace/prettier-config": "workspace:*",
