+++ packages/sde-utils/src/index.ts
@@ -1,3 +1,5 @@
 export * from "./constants.js";
+export * from "./download.js";
 export * from "./sde.js";
 export * from "./sources/sde.js";
+export * from "./zip.js";
