+++ packages/utils/package.json
@@ -20,11 +20,8 @@
     "@jitaspace/prettier-config": "workspace:*",
     "@jitaspace/tsconfig": "workspace:*",
     "@types/node": "^24.12.4",
-    "@types/node-fetch": "^2.6.12",
     "@types/react": "^19.1.6",
     "eslint": "^9.39.4",
-    "node-fetch": "^2.7.0",
-    "node-stream-zip": "^1.15.0",
     "prettier": "^3.8.3",
     "typescript": "^5.9.3"
   },
