+++ packages/utils/src/index.ts
@@ -1,4 +1,3 @@
-export * from "./download";
 export * from "./esi";
 export * from "./fitting";
 export * from "./math";
@@ -8,4 +7,3 @@ export * from "./sde";
 export * from "./skills";
 export * from "./types";
 export * from "./ui";
-export * from "./zip";
