+++ sonar-project.properties
@@ -8,9 +8,7 @@ sonar.tests=apps/web,packages
 sonar.test.inclusions=**/*.test.ts,**/*.test.tsx,**/*.spec.ts,**/*.spec.tsx
 
 sonar.coverage.exclusions=\
-  packages/sde-utils/**,\
-  packages/utils/src/download.ts,\
-  packages/utils/src/zip.ts
+  packages/sde-utils/**
 
 sonar.javascript.lcov.reportPaths=\
   apps/web/coverage/lcov.info,\
